requirements.txt
langchain 
langchain-community
langchain-openai
langchain
python-dotenv
pypdf
sentence-transformers
faiss-cpu